make restart matlab cross platform safe
Former-commit-id: e9d0020d7703d4603e1808a39731e1d0ea57ad5a [formerly 20d6fa3d373e51a575a4c4397a80692f78337804]
Former-commit-id: 6185d699ee0c08dbe63b47441bb4c134167eb820

+++ dynareOBC/setup/RestartMatlab.m
@@ -1,8 +1,18 @@
 function RestartMatlab( CurrentPath, InputFileName, varargin )
-    fprintf( 1, '\n' );
-    disp( 'dynareOBC needs to restart MATLAB. dynareOBC will attempt to continue after MATLAB is restarted.' );
-    fprintf( 1, '\n' );
-    input( 'Press return to continue, or Ctrl+C to cancel.' );
-    system( [ 'start matlab.exe -sd "' CurrentPath '" -r "try;matlabrc;catch;end;try;startup;catch;end; dynareOBC ' InputFileName ' ' strjoin( varargin ) '"' ] );
-    system( [ 'taskkill /f /t /pid ' num2str( feature( 'getpid' ) ) ] ); 
+    Architecture = computer;
+    if length( Architecture ) >= 5 && strcmp( Architecture( 1:5 ), 'PCWIN' )
+
+        fprintf( 1, '\n' );
+        disp( 'dynareOBC needs to restart MATLAB. dynareOBC will attempt to continue after MATLAB is restarted.' );
+        fprintf( 1, '\n' );
+        input( 'Press return to continue, or Ctrl+C to cancel.' );
+        system( [ 'start matlab.exe -sd "' CurrentPath '" -r "try;matlabrc;catch;end;try;startup;catch;end; dynareOBC ' InputFileName ' ' strjoin( varargin ) '"' ] );
+        system( [ 'taskkill /f /t /pid ' num2str( feature( 'getpid' ) ) ] ); 
+        
+    else
+        
+        disp( 'Please manually restart MATLAB.' );
+        error( 'dynareOBC:RestartMATLAB', 'Please manually restart MATLAB.' );
+    
+    end
 end
